alumni-portal/backend/requirements.txt
flask==3.1.1
flask-cors==5.0.1
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.7.1
pymysql==1.1.1
cryptography==44.0.3
python-dotenv==1.1.0
werkzeug==3.1.3
gunicorn==23.0.0
